shift proxy_settings to pipe_views, call it after creating
When in commands, too much of pipe_views leaked through.

Also, bump up a proxy_settings call from the callback on_view_created to
create_destination.
