+++ commands.py
@@ -1,25 +1,11 @@
 import sublime, sublime_plugin
 
 if sublime.version().startswith('3'):
-    from .pipe_views import PipeViews
+    from .pipe_views import PipeViews, proxy_settings
     from . import settings as settings_bv
 else:
-    from pipe_views import PipeViews
+    from pipe_views import PipeViews, proxy_settings
     import settings as settings_bv
-
-def set_settings_listener(receiver, r_key, settings, s_key):
-    settings.clear_on_change(s_key)
-    def callback(*args):
-        val = settings.get(s_key)
-        setattr(receiver, r_key, val)
-    settings.add_on_change(s_key, callback)
-
-def proxy_settings(pipe, settings):
-    # scrolling
-    set_settings_listener(pipe, "scroll_setting", settings, "bv_scroll")
-
-    # enabled
-    set_settings_listener(pipe, "enabled_setting", settings, "bv_enabled")
 
 
 class PlacementPolicy1(object):
@@ -74,8 +60,6 @@ class Pipe(PlacementPolicy1, PipeViews):
     dest_view_name = "Build output"
 
     def on_view_created(self, window, view, pipe):
-        proxy_settings(pipe, view.settings())
-
         group_index, view_index = self.choose_group(window, self.view_launched_build)
         window.set_view_index(view, group_index, view_index)
